requirements.txt
streamlit
biopython
pandas
fastapi
uvicorn
